+++ CHANGELOG.md
@@ -55,6 +55,7 @@ Currently, we plan to implement the following features in the future (in no part
 - 🐘 Added a `logfile()` method to the Postgres interface to retrieve the log file of a local Postgres server
 - 🐘 The parallel query executor now supports optional callbacks
 - Added a `last(n)` method to workloads to retrieve the last _n_ queries
+- Added support for the [JOB-Complex workload](https://www.vldb.org/2025/Workshops/VLDB-Workshops-2025/AIDB/AIDB25_8.pdf)
 
 ### 💀 Breaking changes
 - Eliminated all duplicate imports in the public API. Core classes and functions are now only available from a single location
@@ -70,7 +71,6 @@ Currently, we plan to implement the following features in the future (in no part
   the error is logged to the output file and the execution continues with the next query. This behavior can be controlled
   using the `error_action` parameter.
 - Starting a local Postgres server now supports specifying a log file
-- Much improved Postgres-style plan enumeration which should work better around hinting-related issues
 - When the textbook pipeline auto-selects a Postgres-style enumerator, the enumerator is now configured similarly to the target
   database. E.g., if the textbook is setup for a Postgres database that has nested-loop joins disabled, the enumerator will
   also disable these joins.